corrected warning
git-svn-id: svn+ssh://rzstud.rz.uni-karlsruhe.de/home/ws/uhcnl/repos/Thesis/trunk/src/libs/libcalexxx@710 31e22897-37e7-4591-b389-d3fa21440f69

+++ parser.cc
@@ -89,7 +89,7 @@ namespace calex
         // check if obligatory system parameter
         if (0 == id.size())
         {
-          int sep = str.find("|");
+          size_t sep = str.find("|");
           CALEX_assert(std::string::npos != sep, "Invalid parameter passed.");
           id = str.substr(0,sep);
           iss.str(str.substr(sep+1));
@@ -105,7 +105,7 @@ namespace calex
         // check if obligatory system parameter
         if (0 == id.size())
         {
-          int sep = str.find("|");
+          size_t sep = str.find("|");
           CALEX_assert(std::string::npos != sep, "Invalid parameter passed.");
           id = str.substr(0,sep);
           iss.str(str.substr(sep+1));
